update: modal with ticket detail part
